parser: Split parser into two sections to allow cleaner importing. Also copied robocop.push to avoid importing that module.
(cherry picked from commit 1a31362ecb37353ea36352b230573be0f6058f68)
